+++ src/tile.h
@@ -36,10 +36,10 @@ public:
 	inline void id(uint8_t id) { _id = id; }
 	inline bool undefined(void) const { return _undefined; }
 	inline void undefined(bool u) { _undefined = u; }
-	inline const uchar *rgb(Palette p) const { return _undefined ? _undefined_rgb : _rgb[p]; }
+	inline const uchar *rgb(Palette p) const { return _undefined ? _undefined_rgb : _rgb[(int)p]; }
 	inline Hue hue(int x, int y) const { return _hues[y * TILE_SIZE + x]; }
-	inline uchar *colored_pixel(Palette p, int x, int y) { return &_rgb[p][TILE_PIXEL_OFFSET(x, y)]; }
-	inline const uchar *const_colored_pixel(Palette p, int x, int y) const { return &_rgb[p][TILE_PIXEL_OFFSET(x, y)]; }
+	inline uchar *colored_pixel(Palette p, int x, int y) { return &_rgb[(int)p][TILE_PIXEL_OFFSET(x, y)]; }
+	inline const uchar *const_colored_pixel(Palette p, int x, int y) const { return &_rgb[(int)p][TILE_PIXEL_OFFSET(x, y)]; }
 	inline uchar *monochrome_pixel(int x, int y) { return _monochrome_rgb + TILE_PIXEL_OFFSET(x, y); }
 	inline const uchar *const_monochrome_pixel(int x, int y) const { return _monochrome_rgb + TILE_PIXEL_OFFSET(x, y); }
 	inline uchar *undefined_pixel(int x, int y) { return _undefined_rgb + TILE_PIXEL_OFFSET(x, y); }
